Extract PTS pacing into its own module (src/pacing.rs)
